+++ R/gg_day.r
@@ -276,6 +276,9 @@ gg_day <- function(dataset,
   
   # Return --------------------------------------------------------------
   if(interactive) {
+    if (!requireNamespace("plotly", quietly = TRUE)) {
+      stop("Package 'plotly' is required for interactive use. Please install it.", call. = FALSE)
+    }
     Plot %>% plotly::ggplotly()
   }
   else Plot