+++ R/gg_days.R
@@ -196,6 +196,9 @@ gg_days <- function(dataset,
   
   # Return --------------------------------------------------------------
   if(interactive) {
+    if (!requireNamespace("plotly", quietly = TRUE)) {
+      stop("Package 'plotly' is required for interactive use. Please install it.", call. = FALSE)
+    }
     Plot %>% plotly::ggplotly()
   }
   else Plot