+++ R/import_LL.R
@@ -51,7 +51,10 @@
 #'   column. If the column is not present it will add this column and fill it
 #'   with the filename of the importfile (see param `auto.id`).
 #' * `print_n` can be used if you want to see more rows from the observation intervals
-#' * `remove_duplicates` can be used if identical observations are present within or across multiple files. The default is `FALSE`. The function keeps only unique observations (=rows) if set to' TRUE'. This is a convenience implementation of [dplyr::distinct()].
+#' * `remove_duplicates` can be used if identical observations are present 
+#'   within or across multiple files. The default is `FALSE`. The function keeps 
+#'   only unique observations (=rows) if set to' TRUE'. This is a convenience 
+#'   implementation of [dplyr::distinct()].
 #'
 #' @param ... Parameters that get handed down to the specific import functions
 #' @param device From what device do you want to import? For a few devices,
@@ -413,9 +416,18 @@ imports <- function(device,
              which(table(data$Id) < 2) %>% names()
         )
       }
-      
+      # browser()
       #if there are duplicate rows, remove them and print an info message
-      duplicates <- suppressMessages(janitor::get_dupes(data, -file.name) %>% nrow())
+      var_names <- names(data) |> setdiff("file.name")
+      duplicate.table <- 
+        data |>  
+        dplyr::add_count(dplyr::pick(dplyr::all_of(var_names)), name = "dupes") |> 
+        dplyr::filter(dupes > 1) |> 
+        dplyr::arrange(dplyr::desc(dupes))
+      duplicates <- 
+        duplicate.table |> 
+        nrow()
+      
       orig_rows <- data %>% nrow()
       
       if(duplicates > 0 & remove_duplicates) {
@@ -425,9 +437,9 @@ imports <- function(device,
       
       #if there are untreated duplicate rows, give a warning
       if(duplicates > 0 & !remove_duplicates) {
-        messages <- paste0(format(duplicates, big.mark = "'"), " rows in your dataset(s) are identical to at least one other row. This causes problems during analysis. Please set `remove_duplicates = TRUE` during import. Import will be stopped now and a dataframe with the duplicate rows returned \nIf you want to find out which entries are duplicates. Use `{replace_with_data_object} %>% janitor::get_dupes(-file.name) on your imported dataset.\n")
+        messages <- paste0(format(duplicates, big.mark = "'"), " rows in your dataset(s) are identical to at least one other row. This causes problems during analysis. Please set `remove_duplicates = TRUE` during import. Import will be stopped now and a dataframe with the duplicate rows returned.\n")
         warning(messages)
-        return(janitor::get_dupes(data, -file.name))
+        return(duplicate.table)
       }
       
       #if dst_adjustment is TRUE, adjust the datetime column
@@ -497,9 +509,9 @@ import <- purrr::imap(import_expr, \(x, idx) imports(idx,x))
 #' #the new one is identical to the old one in terms of the function body
 #' identical(body(import$ActLumus), body(new_import$ActLumus))
 #'
-#' #change the import expression for the LYS device to add a message at the top
+#' #change the import expression for the ActLumus device to add a message at the top
 #' new_import_expr <- ll_import_expr()
-#' new_import_expr$ActLumus[[4]] <-
+#' new_import_expr$ActLumus[[6]] <-
 #' rlang::expr({ cat("**This is a new import function**\n")
 #' data
 #' })